requirements.txt
# Ultimate Voice Assistant Requirements for Windows

# IMPORTANT: Install PyTorch first (CUDA/GPU by default) from the official index.
#   GPU (CUDA 12.4 example):
#       pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
#   CPU (only if you do not have a supported GPU):
#       pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# Core ML/AI Libraries
transformers
langchain>=0.3
langchain-community
langchain-core
accelerate>=0.31.0

# ASR (Speech Recognition)
coqui-whisper
faster-whisper>=1.1.0

# Audio I/O & VAD
pyaudio
webrtcvad
sounddevice>=0.5.0
soundfile>=0.12.1
numpy

# TTS - Kokoro (loads via torch.hub in voice-assistant)
kokoro>=0.9.2

# LLM backends & clients
openai>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.27.0

# Configuration
pyyaml

# Utilities
rich>=13.0.0

# System-wide hotkeys and text injection (Windows)
pynput
pywin32

# Screenshot support
pyautogui
Pillow

# Wake word
pvporcupine>=3.0.0
